thumbnailer/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
-e file:../pycommon#egg=pycommon
aiofiles==0.4.0           # via sanic
aiohttp==3.6.2            # via elasticsearch-async
async-timeout==3.0.1      # via aiohttp, elasticsearch-async
attrs==19.3.0             # via aiohttp
certifi==2019.11.28       # via httpx
chardet==3.0.4            # via aiohttp, httpx
dnspython==1.16.0
elasticsearch-async==6.2.0
elasticsearch==7.1.0      # via elasticsearch-async
h11==0.8.1                # via httpx
h2==3.1.1                 # via httpx
hpack==3.0.0              # via h2
hstspreload==2020.1.7     # via httpx
httptools==0.0.13         # via sanic
httpx==0.9.3              # via sanic
hyperframe==5.2.0         # via h2
idna==2.7                 # via httpx, yarl
motor==2.0.0
multidict==4.7.4          # via aiohttp, sanic, yarl
mypy-extensions==0.4.3    # via mypy
mypy==0.740
pillow==7.0.0
pymongo==3.7.1
python-bsonjs==0.2.0
python-dateutil==2.7.5
rfc3986==1.3.2            # via httpx
sanic==19.12.2
six==1.13.0               # via python-dateutil
sniffio==1.1.0            # via httpx
typed-ast==1.4.0          # via mypy
typing-extensions==3.7.4.1  # via mypy
ujson==1.35               # via sanic
urllib3==1.25.7           # via elasticsearch
uvloop==0.14.0            # via sanic
wand==0.5.8
websockets==8.1           # via sanic
yarl==1.4.2               # via aiohttp
